DjangoBackend/requirements.txt
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cssselect==1.1.0
defusedxml==0.7.1
Django==3.2.7
django-cors-headers==3.8.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
djoser==2.1.0
gunicorn==20.0.4
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
lxml==4.6.3
MarkupSafe==2.0.1
oauthlib==3.1.1
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyfiglet==0.8.post1
PyJWT==2.1.0
pymongo==3.12.0
Pygtail
pyOpenSSL==20.0.1
python-scrapyd-api==2.1.2
python3-openid==3.2.0
pytz==2021.1
queuelib==1.6.2
requests==2.26.0
requests-oauthlib==1.3.0
Scrapy==2.5.0
scrapy-user-agents==0.1.1
scrapyd==1.2.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.2.4
termcolor==1.1.0
Twisted==21.7.0
typing-extensions==3.10.0.2
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.6
user-agents==2.2.0
w3lib==1.22.0
zope.interface==5.4.0
